+++ src/graph/common.jl
@@ -1,15 +1,16 @@
 # TODO:
-# - add keyword argument to choose whether to enable/disable showing indices?
-# - parse regulators, throttle control valves, others???
+# - use keyword argument to enable/disable showing indices?
+# - parse the "status" flag for every node and edge? needed for `des_pipes` (design
+#   problem); not sure of other use cases
 # * for "epanet-data" objects, identify shutoff valves; might want to do that in
 #   WaterModels?
-# - use PyPDF2 to merge colorbar (will remove ghostscript dependency)
 
 
 
 """
 Build pygraphviz graph object from a WaterModels network dictionary parsed from an EPANET
-file.
+file. If a `solution` dict is provided, it should be for the same time as that of the `data`
+dict.
 """
 function build_graph(data::Dict{String, <:Any},
                      solution::Union{Nothing, Dict{String, <:Any}} = nothing)
@@ -125,8 +126,10 @@ function node_labels!(G::PyCall.PyObject, nodes::Dict{String,Any})
         
         if node_type == "reservoir"
             PyCall.set!(nodeobj.attr, "label", "Rsvr\n"*label)
+            PyCall.set!(nodeobj.attr, "shape", "diamond")
         elseif node_type == "tank"
             PyCall.set!(nodeobj.attr, "label", "Tank\n"*label)
+            PyCall.set!(nodeobj.attr, "shape", "rectangle")
         else
             dem = node["flow_nominal"] 
 
@@ -171,10 +174,18 @@ end
 function _add_pipe_to_graph!(graph::PyCall.PyObject, pipe::Dict{String, <:Any}, max_diameter::Float64)
     index, label = string(pipe["index"]), _get_comp_label(pipe)
     penwidth = pipe["diameter"] * inv(max_diameter) * 10.0
+    if penwidth > 8.0
+        pad = "  "
+    elseif penwidth > 4.0
+        pad = " "
+    else
+        pad = ""
+    end
     dir, arrowhead = _get_link_dir(pipe), _get_link_arrowhead(pipe)
     length = @sprintf("%.5g m", pipe["length"])
     graph.add_edge(pipe["node_fr"], pipe["node_to"], index, dir = dir, headclip = "true",
-        arrowhead = arrowhead, penwidth = penwidth, label = "$(label)\n$(length)")
+                   arrowhead = arrowhead, penwidth = penwidth,
+                   label = "$(pad)$(label)\n$(pad)$(length)")
 end
 
 
@@ -203,15 +214,20 @@ function _add_regulator_to_graph!(graph::PyCall.PyObject, regulator::Dict{String
     index, label = string(regulator["index"]), _get_comp_label(regulator)
     dir, arrowhead = _get_link_dir(regulator), _get_link_arrowhead(regulator)
     graph.add_edge(regulator["node_fr"], regulator["node_to"], index, dir = dir,
-        arrowhead = arrowhead, label = "Reg\n$(label)")
+        arrowhead = arrowhead, label = "Reg\n$(label)", color = "purple", style = "bold")
 end
 
 
 function _add_valve_to_graph!(graph::PyCall.PyObject, valve::Dict{String, <:Any})
     index, label = string(valve["index"]), _get_comp_label(valve)
     dir, arrowhead = _get_link_dir(valve), _get_link_arrowhead(valve)
+    if dir != "none"
+        label = "CV\n$(label)"
+    else
+        label = "SV\n$(label)"
+    end
     graph.add_edge(valve["node_fr"], valve["node_to"], index, dir = dir,
-        arrowhead = arrowhead, label = "Vlv\n$(label)", color = "blue", style = "bold")
+        arrowhead = arrowhead, label = label, color = "blue", style = "bold")
 end
 
 
@@ -228,9 +244,10 @@ function add_solution!(G::PyCall.PyObject, data::Dict{String,Any},
         PyCall.set!(nodeobj.attr, "label", label*"\nh: "*head)
     end
     # add flow to the labels for pipes and valves
-    pipesplus = ["pipe", "des_pipe", "pump", "regulator", "short_pipe", "valve"]
-
-    for linktype in pipesplus
+    # Byron added this set, but not all of these fields exist in the solution object
+    #pipesplus = ["pipe", "des_pipe", "pump", "regulator", "short_pipe", "valve"]
+    links = ["pump", "pipe", "valve"]
+    for linktype in links
         for (key,pipesol) in solution[linktype]
             flow = _val_string_cut(pipesol["q"], 1e-10)
             link = data[linktype][key]
@@ -239,19 +256,6 @@ function add_solution!(G::PyCall.PyObject, data::Dict{String,Any},
             label = get(edgeobj.attr, "label")
             PyCall.set!(edgeobj.attr, "label", label*"\nq: "*flow)
         end
-    end
-    # add flow and gain to the pump labels
-    ### remove the gain and only show flow?? if show gain for pumps, logically should show
-    ### headloss for the other links.... also, would only need one loop block if only show
-    ### flow, JJS 12/29/20
-    for (key,pumpsol) in solution["pump"]
-        flow = _val_string_cut(pumpsol["q"], 1e-10)
-        gain = _val_string_cut(pumpsol["g"], 1e-3)
-        pump = data["pump"][key]
-        # may also need to use `key` if multiple pumps between nodes 
-        edgeobj = @pycall G."get_edge"(pump["node_fr"], pump["node_to"])::PyObject 
-        label = get(edgeobj.attr, "label")
-        PyCall.set!(edgeobj.attr, "label", label*"\nq: "*flow*"\ng: "*gain)
     end
 end
 
@@ -270,28 +274,31 @@ Write out to a file a visualization for a WaterModels network dictionary parsed 
 EPANET file. `basefilename` should not include an extension and will be appended with
 `_w_cb.pdf` in the final output file, which is a multi-page PDF. The `layout` option equates
 to the layout functions of graphviz (dot, neato, etc.). Use `del_files=false` to keep the
-intermediate files.
+intermediate files. If a `solution` dict is provided, it should be for the same time as that
+of the `data` dict.
 """
 function write_visualization(data::Dict{String,Any}, basefilename::String,
                              solution::Union{Nothing, Dict{String,Any}}=nothing;
-                             layout::String="dot", del_files::Bool=true)
+                             layout::String="dot", sep_page::Bool=false,
+                             del_files::Bool=true)
     # TODO:
-    # - pass through arguments to `write_graph`
+    # - pass through general graphviz arguments to `write_graph`
 
     #gvfile = basefilename*".gv"
-    pdffile = basefilename*".pdf"
+    gpdffile = basefilename*"_graph.pdf"
     cbfile = basefilename*"_cbar.pdf"
-    outfile = basefilename*"_w_cb.pdf"
+    outfile = basefilename*"_graph_w_cb.pdf"
     
     G = build_graph(data, solution)
-    write_graph(G, pdffile, layout)
+    write_graph(G, gpdffile, layout)
     colorbar(G, cbfile)
 
-    # add option -dAutoRotatePages=/None ?
-    run(`gs -sDEVICE=pdfwrite -dNOPAUSE -dQUIET -dBATCH -sOutputFile=$outfile $cbfile $pdffile`)
+    # note that `stack_bar` is a python function from `wntr_vis.py`
+    stack_cbar(gpdffile, cbfile, outfile, sep_page)
+    
     # delete the intermediate files
     if del_files
-        run(`rm $pdffile $cbfile`)
+        run(`rm $gpdffile $cbfile`)
     end
 end
 
@@ -325,6 +332,8 @@ function colorbar(G::PyCall.PyObject, filename::String)
     x = reshape(collect(range(0.0, stop=1.0, length=100)), (1,:))
     Plots.heatmap(x, c=:viridis, size=(500,100), legend=:none, yaxis=false)
     Plots.plot!(xticks=(0:50:100, [elmin, elmid, elmax]))
+    Plots.plot!(yticks=false) # a regression requires this for GR, JJS 1/4/21,
+                              # https://github.com/JuliaPlots/Plots.jl/issues/3019
     Plots.title!("Elevation")
     Plots.savefig(filename)
 end