+++ CHANGELOG.md
@@ -4,6 +4,8 @@ WaterModelsAnalytics.jl Change Log
 ### v0.3.0
 - Updates for WaterModels v0.6.
 - Implementation of WNTR simulation of WaterModels data.
+- Use PyPDF2 rather than ghostscript to merge colorbar and graph
+- Updates to visualizations (different node shapes, pipe thickness, and more)
 
 ### v0.2.0
 - Updates for WaterModels v0.4.